Refactored Components to use React Image  Components and removed unused import
