+++ app/(ui)/components/SingleCorner.tsx
@@ -1,6 +1,8 @@
+import Image from "next/image";
+
 export default function SingleCorner({ size = 50, rotation = 0 }) {
   return (
-    <img
+    <Image
       src="/assets/NormalCorner.svg"
       alt=""
       width={size}
